tests: hooks: xfail test on fc28
One of the tests from the 'hooks_test' module fails in fc28 x travis
environment. The reason behind the failure is unknown - attempts
to reproduce it manually have failed. Fc28 is EOLed, but is used by the
storage team, because currently it's the only distro we have that offers
py36. Since running the test manually on rhel8 (also on py36) succeeds,
we're xfailing it only for this combination for now.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>