py3: Enable backtrace tests for 'py-watch' on Fedora
Back in the days of fc27 there was an issue [1] when trying to get
a python backtrace in gdb. Things got fixed, so we should re-enable this
test.

Here's an example output for a script like the one we have embedded
in 'pywatch_test':

 (gdb) py-bt
 Traceback (most recent call first):
   <built-in method sleep of module object at remote 0x7fe9e61c4fb0>
   File "sss.py", line 8, in inner
     time.sleep(100000)
   File "sss.py", line 5, in outer
     inner()
   File "sss.py", line 10, in <module>
     outer()

Since the output seems to be a bit different than the one we expect
(no braces after function name), the assertions are aligned to current
output.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1557069

Change-Id: I4e8e211702cf3844f366fba7287b25f0c7ad3dca
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>