tool: libvirt: Dynamic socket requirements
Since libvirt switched to socket activation, one needs to enable
its socket units to be able to connect with TCP/TLS.

Until now, we simply added the requirements to socket units
to supervdsm's service definition. This approach however had two flaws:
- we required both TCP and TLS sockets at all times
- it didn't guarantee that the socket units will be up before
  'libvirtd.service' and socket units will fail to start if the daemon
  is already running

This patch changes the approach to a more dynamic one. During
reconfiguration of libvirt module with 'vdsm-tool', depending on the
configuration of vdsm (ssl/non-ssl), we inject a requirement to the
appropriate socket unit on the 'libvirtd.service' itself with systemd.
This is done by creating a symlink like:

 /etc/systemd/system/libvirtd.service.requires/libvirtd-tls.socket --> /usr/lib/systemd/system/libvirtd-tls.socket

This both gives us the guarantee, that during next start
of 'libvirtd.service' the socket will also be up, and makes only
TCP or TLS socket required.

During deconfiguration we remove the symlinks.

Since tests for all 'vdsm-tool' configurators are still written in nose,
a significant amount of time will be needed to adjust them to the new
way of working. This patch disables all currently failing tests and they
will be reenabled later.

Change-Id: Ia602614cf7736197fc5834842925e4a49c7aa64a
Bug-Url: https://bugzilla.redhat.com/1818554
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>