tool: libvirt: Don't shutdown libvirtd on timeout
By default 'libvirtd' starts with '--timeout 120' flag. The effect
of this flag is that after 120 secs, if no client has made a connection
to the daemon, it will shut itself down.

This is the essence of socket activation feature. However given the
problematic fact that libvirt's socket units cannot be started after
the daemon is up and multiple scenarios where we juggle the services
(i.e. host deployment process), it's better if we keep things simple
and let 'libvirtd' stay up once it's up.

Change-Id: I114e046131c1e9edb36b59e16a8e94b8e30ecfa2
Bug-Url: https://bugzilla.redhat.com/1818554
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>